01-docker-terraform/requirements.txt
sqlalchemy
pandas
python-dotenv
psycopg2